refactor existing hello world block into its own folder
